fix: api перенести хендлер HTML в другой файл
