+++ internal/server/api/state_handlers.go
@@ -2,6 +2,7 @@ package api
 
 import (
 	"net/http"
+	"text/template"
 )
 
 // PingStore создает хендлер, который пингует хранилище. Если хранилище может связаться с базой данных,
@@ -15,4 +16,54 @@ func PingStore(store Operator) http.HandlerFunc {
 		}
 	}
 
+}
+
+// DisplayHTML создает хендлер HTTP запроса. Этот хендлер формирует простую
+// HTML страничку, где указаны все известные на данный момент метрики
+func DisplayHTML(op Operator) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		SetContentType(w, TypeTextHTML)
+
+		htmlTemplate := `
+	{{ if .ListCounters}}
+		Counters
+			<ul> 
+				{{range .ListCounters -}}
+					<li>{{.}}</li>
+				{{ end }}
+			</ul>
+	{{ else }}
+		<p>No Counters</p>
+	{{end}}
+	{{ if .ListGauges}}
+		Gauges
+			<ul> 
+				{{range .ListGauges -}}
+					<li>{{.}}</li>
+				{{ end }}
+			</ul>
+	{{ else }}
+		<p>No Gauges</p>
+	{{end}}
+	`
+
+		tmpl, err := template.New("simple").Parse(htmlTemplate)
+		if err != nil {
+			HTTPErrorWithLogging(w, http.StatusInternalServerError, "Не удалось пропарсить HTML шаблон: %v", err)
+			return
+		}
+		cs, gs, err := op.List(r.Context())
+		if err != nil {
+			HTTPErrorWithLogging(w, http.StatusInternalServerError, "Ошибка получения списка метрик из хранилища: %v", err)
+			return
+		}
+		err = tmpl.Execute(w, struct {
+			ListCounters []string
+			ListGauges   []string
+		}{ListCounters: cs, ListGauges: gs})
+		if err != nil {
+			HTTPErrorWithLogging(w, http.StatusInternalServerError, "Ошибка запуска шаблона HTML: %v", err)
+			return
+		}
+	}
 }
